Add LogsSettings into LogsVmd. Refactoring LogsPanel infrastructure
