+++ AvaloniaUIClient/Infrastructure/Views/DevPanelViews/LogsPanel/LogsSettingsView.axaml.cs
@@ -1,0 +1,17 @@
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace AvaloniaUIClient.Infrastructure.Views.DevPanelViews.LogsPanel;
+
+public partial class LogsSettingsView : UserControl
+{
+    public LogsSettingsView()
+    {
+        InitializeComponent();
+    }
+
+    private void InitializeComponent()
+    {
+        AvaloniaXamlLoader.Load(this);
+    }
+}
