+++ Core/Infrastructure/IOC/StoresRegistration.cs
@@ -6,6 +6,7 @@ using Core.Infrastructure.Models.SettingsModels;
 using Core.Infrastructure.Stores;
 using Core.Infrastructure.VMD;
 using Core.Stores;
+using Core.VMD.DevPanelVmds;
 using Microsoft.Extensions.DependencyInjection;
 using Serilog.Events;
 
@@ -23,6 +24,7 @@ public partial class IocRegistration
         services
             .AddSingleton<IStore<ITitleVmd>, TitleVmdStore>()
             .AddSingleton<IStore<Settings>, SettingsStore>()
+            .AddTransient<LogsSettingsVmd>()
             .AddSingleton<BaseTimerReactiveStore<Settings>>(s=>(BaseTimerReactiveStore<Settings>)s.GetService<IStore<Settings>>());
 
     private static IServiceCollection InfrStoresRegs(this IServiceCollection services) =>
